#959 Rename Load Model .cc files from extended to nompi
